# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catat Meter API V1",
        "description": "API documentation for Catat Meter - Water Utility Management System. This API provides endpoints for managing customers, meter readings, billing, and payments for water utility companies (PAM).",
        "contact": {
            "email": "support@catatmeter.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Catat Meter API Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/customers/{customerId}/bills/pay": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Pay customer bills",
                "description": "Process payment for one or more customer bills",
                "operationId": "payBills",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bill_ids",
                                    "payment_method"
                                ],
                                "properties": {
                                    "bill_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "transfer",
                                            "ewallet"
                                        ],
                                        "example": "cash"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "1 bills paid successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "customer_name": {
                                                    "type": "string",
                                                    "example": "Fahmi"
                                                },
                                                "total_payment": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 55000
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "example": "2025-11-18 17:58:00"
                                                },
                                                "updated_bills": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "bill_number": {
                                                                "type": "string",
                                                                "example":"BILL-1-202511-0135"
                                                            },
                                                            "total_bill": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 6600
                                                            }
                                                        }
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}